renamed workflow to aiq_agno_personal_finance
