+++ examples/agno_personal_finance/Dockerfile
@@ -38,17 +38,17 @@ RUN --mount=type=cache,id=uv_cache,target=/root/.cache/uv,sharing=locked \
     export SETUPTOOLS_SCM_PRETEND_VERSION_FOR_AGENTIQ=${AIQ_VERSION} && \
     export SETUPTOOLS_SCM_PRETEND_VERSION_FOR_AGENTIQ_TEST=${AIQ_VERSION} && \
     export SETUPTOOLS_SCM_PRETEND_VERSION_FOR_AGENTIQ_AGNO=${AIQ_VERSION} && \
-    export SETUPTOOLS_SCM_PRETEND_VERSION_FOR_AGNO_PERSONAL_FINANCE=${AIQ_VERSION} && \
+    export SETUPTOOLS_SCM_PRETEND_VERSION_FOR_AIQ_AGNO_PERSONAL_FINANCE=${AIQ_VERSION} && \
     uv venv --python ${PYTHON_VERSION} /workspace/.venv && \
     uv sync --link-mode=copy --compile-bytecode --python ${PYTHON_VERSION} && \
     uv pip install -e '.[agno]' && \
     uv pip install --link-mode=copy ./examples/agno_personal_finance 
 
 # Set the config file environment variable
-ENV AIQ_CONFIG_FILE=/workspace/examples/agno_personal_finance/src/agno_personal_finance/configs/config.yml
+ENV AIQ_CONFIG_FILE=/workspace/examples/agno_personal_finance/src/aiq_agno_personal_finance/configs/config.yml
 
 # Enivronment variables for the venv
 ENV PATH="/workspace/.venv/bin:$PATH"
 
 # Define the entry point to start the server
-ENTRYPOINT ["aiq", "serve", "--config_file=/workspace/examples/agno_personal_finance/src/agno_personal_finance/configs/config.yml", "--host", "0.0.0.0"]
+ENTRYPOINT ["aiq", "serve", "--config_file=/workspace/examples/agno_personal_finance/src/aiq_agno_personal_finance/configs/config.yml", "--host", "0.0.0.0"]
