+++ examples/agno_personal_finance/pyproject.toml
@@ -6,7 +6,7 @@ requires = ["setuptools >= 64", "setuptools-scm>=8"]
 root = "../.."
 
 [project]
-name = "agno_personal_finance"
+name = "aiq_agno_personal_finance"
 dynamic = ["version"]
 dependencies = [
   "agentiq[agno]",
@@ -21,4 +21,4 @@ classifiers = ["Programming Language :: Python"]
 agentiq = { path = "../..", editable = true }
 
 [project.entry-points.'aiq.components']
-agno_personal_finance = "agno_personal_finance.register"
+aiq_agno_personal_finance = "aiq_agno_personal_finance.register"
