+++ examples/agno_personal_finance/src/aiq_agno_personal_finance/agno_personal_finance_function.py
