+++ examples/agno_personal_finance/src/aiq_agno_personal_finance/register.py
